feat: dropdown red color

+++ packages/server/src/modules/posts/posts.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, Injectable } from '@nestjs/common';
 import { CreatePostDto } from './dto/create-post.dto';
 import { PostsRepository } from './posts.repository';
 
@@ -10,14 +10,20 @@ export class PostsService {
   }
 
   async getPostById(id: string) {
-    return await this.postRepository.findPostById(id);
+    const post = await this.postRepository.findPostById(id);
+    if (!post) throw new HttpException('Post not found', 404);
+    return post;
   }
 
   async createPost(dto: CreatePostDto, authorId: string) {
     return await this.postRepository.createPost(dto, authorId);
   }
 
-  async deletePost(id: string) {
+  async deletePost(id: string, authorId: string) {
+    const post = await this.postRepository.findPostById(id);
+    if (!post) throw new HttpException('Post not found', 404);
+    if (post.authorId !== authorId)
+      throw new HttpException('You are not the author of this post', 403);
     return await this.postRepository.deletePost(id);
   }
 }
